Added low write script

+++ low_write_phc.py
@@ -19,6 +19,6 @@ f = open('phc_eqs/'+str(script_filename)+'.sh','w')
 for a in range(a1,a2+1):
 	for t in range(reps):
 		for j in range(m):
-			md = math.ceil(math.log(a))
+			md = int(math.ceil(math.log(a)))
 			base_name = 'ln'+str(a)+'md'+str(md)
 			f.write('../.././phc -b '+base_name+'_'+str(t)+'_eqs.txt '+base_name+'_'+str(t)+'_roots'+str(j)+'.txt\n')
